requirements.txt
anyio==4.6.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
coverage==7.6.1
Django==5.1.1
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-django==4.9.0
pytest-mock==3.14.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
urllib3==2.2.3
